+++ mopinion/dataclasses.py
@@ -10,7 +10,6 @@ import re
 __all__ = [
     "Credentials",
     "EndPoint",
-    "ApiRequestArguments",
     "ResourceUri",
     "ResourceVerbosity",
 ]
@@ -37,19 +36,22 @@ class EndPoint(Argument):
 
         # endpoint must be one of these
         regexps = [
-            r"/token$",
-            r"/ping$",
-            r"/account$",
-            r"/deployments$",
-            r"/deployments/\w+$",
-            r"/datasets$",
-            r"/datasets/\d+$",
-            r"/datasets/\d+/fields$",
-            r"/reports/\d+/fields$",
-            r"/datasets/\d+/feedback$",
-            r"/reports/\d+/feedback$",
-            r"/reports/\d+$",
-            r"/reports$",
+            r"^/token$",
+            r"^/ping$",
+            r"^/account$",
+            # deployments
+            r"^/deployments$",
+            r"^/deployments/\w+$",
+            # datasets
+            r"^/datasets$",
+            r"^/datasets/\d+$",
+            r"^/datasets/\d+/fields$",
+            r"^/datasets/\d+/feedback$",
+            # reports
+            r"^/reports$",
+            r"^/reports/\d+$",
+            r"^/reports/\d+/fields$",
+            r"^/reports/\d+/feedback$",
         ]
         regexp = re.compile("|".join(regexps), re.IGNORECASE)
         if not regexp.search(self.path):
@@ -57,7 +59,7 @@ class EndPoint(Argument):
 
 
 @dataclass(frozen=True)
-class ApiRequestArguments(Argument):
+class RequestArguments(Argument):
     endpoint: EndPoint
     content_negotiation: str
     verbosity: str
@@ -65,10 +67,10 @@ class ApiRequestArguments(Argument):
 
     def __post_init__(self):
         # verbosity levels
-        if self.verbosity.lower() not in settings.VERBOSITY_lEVELS:
+        if self.verbosity.lower() not in settings.VERBOSITY_LEVELS:
             raise ValueError(
                 f"'{self.verbosity}' is not a valid verbosity level. Please consider one of: "
-                f"'{', '.join(settings.VERBOSITY_lEVELS)}'"
+                f"'{', '.join(settings.VERBOSITY_LEVELS)}'"
             )
 
         if self.content_negotiation not in settings.CONTENT_NEGOTIATIONS:
@@ -92,32 +94,6 @@ class ResourceUri(Argument):
     sub_resource_name: Optional[str]
 
     def __post_init__(self):
-        from mopinion.client import MopinionClient
-
-        allowed_resources = [
-            MopinionClient.RESOURCE_ACCOUNT,
-            MopinionClient.RESOURCE_DEPLOYMENTS,
-            MopinionClient.RESOURCE_DATASETS,
-            MopinionClient.RESOURCE_REPORTS,
-        ]
-        allowed_subresources = [
-            MopinionClient.SUBRESOURCE_FEEDBACK,
-            MopinionClient.SUBRESOURCE_FIELDS,
-        ]
-
-        if self.resource_name and self.resource_name not in allowed_resources:
-            raise ValueError(
-                f"Resource name {self.resource_name} must be one of {allowed_resources}"
-            )
-
-        if (
-            self.sub_resource_name
-            and self.sub_resource_name not in allowed_subresources
-        ):
-            raise ValueError(
-                f"Subresource name {self.sub_resource_name} must be one of {allowed_subresources}"
-            )
-
         endpoint = f"/{self.resource_name}"
         if self.resource_id:
             endpoint += f"/{self.resource_id}"
@@ -133,11 +109,8 @@ class ResourceVerbosity(Argument):
 
     def __post_init__(self):
         # if we want to iterate we need metadata, verbosity higher or equal than normal
-        if (
-            self.iterator
-            and self.verbosity.lower() not in settings.ITERATE_VERBOSITY_lEVELS
-        ):
+        if self.iterator and self.verbosity.lower() not in ["normal", "full"]:
             raise ValueError(
                 f"'{self.verbosity}' is not a valid verbosity level. Please "
-                f"consider one of: '{', '.join(settings.ITERATE_VERBOSITY_lEVELS)}'"
+                f"consider one of: '{', '.join(['normal', 'full'])}'"
             )