+++ CHANGES.rst
@@ -2,6 +2,21 @@ Changelog
 ========================================================
 
 
+0.0.7 (2022-02-23)
+-------------------
+
+- Fixes in documentation.
+
+- Implement `__enter__()` and `__exit__()` methods.
+
+- Added methods: `get_account`, `get_deployments`,
+  `get_datasets`, `get_datasets_fields`,
+  `get_datasets_feedback`, `get_reports`,
+  `get_reports_fields`, `get_reports_feedback`.
+
+- Increased compatibility from Python 3.6 to 3.10.
+
+
 0.0.6 (2021-02-11)
 -------------------
 
@@ -53,4 +68,3 @@ Changelog
 - Test suite.
 
 - Request examples.
-